+++ poetry/mixology/version_solver.py
@@ -374,6 +374,9 @@ class VersionSolver:
 
             return dependency.name
 
+        if not version.is_root():
+            version = self._provider.complete_package(version)
+
         conflict = False
         for incompatibility in self._provider.incompatibilities_for(version):
             self._add_incompatibility(incompatibility)