+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Fixed `self:update` not picking up new versions.
 - Fixed a `RuntimeError` on Python 3.7.
 - Fixed bad version number being picked with private repositories.
+- Fixed handling of duplicate dependencies with same constraint.
 
 
 ## [0.10.2] - 2018-05-31